% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_folders.R
\name{cleanup_folders}
\alias{cleanup_folders}
\title{Cleanup Folders}
\usage{
cleanup_folders(
  folders = folders::get_folders(),
  conf_file = here::here("folders.yml"),
  keep_conf = TRUE
)
}
\arguments{
\item{folders}{(list) A named list of standard folders for an R project.
(Default: folders::get_folders())}

\item{conf_file}{(character) Configuration file to read/write.
See: config::get(). (Default: here::here("folders.yml"))}

\item{keep_conf}{(boolean) Keep the "folders.yml" file if TRUE. (Default: TRUE)}
}
\value{
(integer) A vector of results: 0 for success; 1 for failure; NULL for skipped.
}
\description{
Remove empty folders from the folders list as well as the configuration file.
}
\section{Details}{

Each non-empty folder in the list of folders will be removed. The configuration  
file will also be removed if keep_conf is set to FALSE.
}

\examples{
\dontrun{
folders <- get_folders()
result <- create_folders(folders)
result <- cleanup_folders(folders)
}
}
\keyword{configuration,}
\keyword{folders,}
\keyword{portability,}
\keyword{reproducibility,}
\keyword{structure}
